% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{add_options}
\alias{add_options}
\title{Chart options}
\usage{
add_options(chart, id = NULL)
}
\arguments{
\item{id}{A character string specifying the id of the chart.}
}
\description{
Update charts options. Currently only used to assign an HTML id to the chart.
}
