% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes.R
\name{add_x_axis}
\alias{add_x_axis}
\alias{add_y_axis}
\title{Configure a chart's x- and y-axis}
\usage{
add_x_axis(chart, title = NULL, labels = TRUE, grid = TRUE, offset = 30,
  position = "end")

add_y_axis(chart, title = NULL, labels = TRUE, grid = TRUE, offset = 40,
  position = "start")
}
\arguments{
\item{chart}{A \code{chartist} chart.}

\item{title}{A character string specifying the axis title, defaults to \code{NULL}
in which case the data column name is used as the axis title.}

\item{labels}{If \code{TRUE}, labels are shown on the x-axis, defaults to \code{TRUE}.}

\item{grid}{If \code{TRUE}, the x-axis grid lines are shown, defaults to \code{TRUE}.}

\item{offset}{A numeric value specifying the offset of the axis labels, the
x-axis offset defaults to \code{30} and the y-axis offset defaults to \code{40}.}

\item{position}{One of \code{"start"} or \code{"end"}, the x-axis position defaults
to \code{"end"} and the y-axis position defaults to \code{"start"}.}
}
\description{
The x-axis and y-axis options.
}
