% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartisan.R
\docType{package}
\name{chartisan}
\alias{chartisan}
\alias{chartisan-package}
\alias{chartisan-package}
\title{chartisan: Chartist.js Charts in R}
\description{
Build chartist.js charts in R. Build
accessible charts.
}
\author{
\strong{Maintainer}: Nathan Teetor \email{nathanteetor@gmail.com}

}
