% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartist.R
\name{chartist}
\alias{chartist}
\title{Build a chartist chart}
\usage{
chartist(data, x, y)
}
\arguments{
\item{data}{A data frame, defaults to \code{NULL}.}

\item{x}{The column}
}
\description{
Start building a chart.
}
\examples{
lines <- iris \%>\%
  chartist(Sepal.Length, Sepal.Width) \%>\%
  layer_line() \%>\%
  layer_ratio_5_8()

lines

points <- iris \%>\%
  chartist(Sepal.Length, Sepal.Width) \%>\%
  layer_point() \%>\%
  layer_ratio_5_8()

points

}
