% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-area.R
\name{layer_area}
\alias{layer_area}
\title{An area chart}
\usage{
layer_area(chart)
}
\arguments{
\item{chart}{A \code{chartist} object.}
}
\description{
Create an area chart.
}
\examples{
areas <- faithful \%>\%
  chartist(waiting, eruptions) \%>\%
  layer_area() \%>\%
  layer_ratio_2_5() \%>\%
  add_x_axis(grid = FALSE)

areas

}
