% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-line.R
\name{layer_line}
\alias{layer_line}
\title{A line chart}
\usage{
layer_line(chart, smooth = FALSE)
}
\arguments{
\item{chart}{A \code{chartist} object.}

\item{smooth}{One of \code{TRUE} or \code{FALSE}, specifying if the line should be
smoothed, defaults to \code{FALSE}.}
}
\description{
Add a line layer to a chart.
}
