% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratios.R
\name{ratios}
\alias{ratios}
\alias{layer_ratio_1_1}
\alias{layer_ratio_3_4}
\alias{layer_ratio_5_8}
\alias{layer_ratio_15_16}
\alias{layer_ratio_8_9}
\alias{layer_ratio_5_6}
\alias{layer_ratio_4_5}
\alias{layer_ratio_2_3}
\alias{layer_ratio_5_8}
\alias{layer_ratio_golden}
\alias{layer_ratio_3_5}
\alias{layer_ratio_9_16}
\alias{layer_ratio_8_15}
\alias{layer_ratio_1_2}
\alias{layer_ratio_2_5}
\alias{layer_ratio_3_8}
\alias{layer_ratio_1_3}
\alias{layer_ratio_1_4}
\title{Chart ratios}
\usage{
layer_ratio_1_1(chart)

layer_ratio_3_4(chart)

layer_ratio_5_8(chart)

layer_ratio_15_16(chart)

layer_ratio_8_9(chart)

layer_ratio_5_6(chart)

layer_ratio_4_5(chart)

layer_ratio_2_3(chart)

layer_ratio_5_8(chart)

layer_ratio_golden(chart)

layer_ratio_3_5(chart)

layer_ratio_9_16(chart)

layer_ratio_8_15(chart)

layer_ratio_1_2(chart)

layer_ratio_2_5(chart)

layer_ratio_3_8(chart)

layer_ratio_1_3(chart)

layer_ratio_1_4(chart)
}
\arguments{
\item{chart}{A \code{chartist} object.}
}
\description{
Change the display ratio of a chart. \code{*_golden} refers to the golden ratio.
}
